+++ NEWS.md
@@ -1,3 +1,8 @@
+PENDING: Dataiter 0.39.1
+========================
+
+* `ListOfDicts.to_data_frame`: Add `strings_as_object` argument
+
 2023-01-21: Dataiter 0.39
 =========================